move language stuff to InfoProvider()
